chore(style): streamline collection expression '[]' + remarks (#427)
This PR streamlines the use of the collection expression `[]` instead of
the old `new()`, and during this streamlining, the summary remarks of
both the MongoDB/NoSQL test fixtures were streamlined as well.

> [!NOTE]
> It is normally not in Arcus' best interest to constantly follow
immediately the newest code syntax as this would decrease the inclusion
of juniors. Though, in some cases like these, the code syntax is
actually an improvement on readability and can be considered a helpful
enhancement.

Closes #426