chore(docs): use pnpm + gh token correctly in auto-bump docs version workflow (#425)
With the recent v2.1 release, there seemed to be a problem with the
auto-bump workflow for the documentation. The GitHub token was not
correctly used and we could improve the workflow by using the PNPM
package manager - which is the recommended way of doing things, and also
stops the re-creation of the `package.lock.json` file, in favor of the
existing `pnpn-lock.yaml` file.