chore(analysis): applied code analysis results in blob storage (#438)
Applies the recommended code analysis to the `.Blob` library. This is
mostly about the logger message templates that can be made as 'source
generated log methods' for improved performance. Though, in our case, it
is also about centralizing the log statements so it is more easily
maintained (setup/teardown in same style); plus now we can reuse even
some log messages.

Partially related to #429