test(integration): remove unnecessary test collections (#476)
We placed certain storage-specific integration tests together, but the
tests are written in such a way that they support parallism (case in
point: the collection attribute was missing from some of them). This was
added previously for a starting point, but with a mature test suite, we
can remove them again.