fix(test.auth): move managed identity as an assembly test fixture (#479)
Due to the full parallelization of the integration tests (see #476), we
need to make sure that there are no shared test fixtures across the
tests. It turns out that we have configured a managed identity
connection for each test context, and since this is pointing to the same
service principal, it could happen that it gets teared down while
another test still needs it.

This PR moves the managed identity test fixture to an assembly test
fixture, so that authentication is configured once and shared across the
tests.